pyproject.toml
[tool.poetry]
name = "ai-meeting-notes-agent"
version = "2.0.0"
description = "AI-powered meeting management platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
websockets = "^12.0"
jinja2 = "^3.1.2"
aiofiles = "^23.2.1"
python-dotenv = "^1.0.0"
google-generativeai = "^0.8.3"
langchain-google-genai = "^1.0.3"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
selenium = "^4.15.2"
webdriver-manager = "^4.0.1"
python-socketio = "^5.10.0"
redis = "^5.0.1"
celery = "^5.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"